# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "EventComment": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "determination": {
            "title": "Determination",
            "type": "string"
          },
          "eventId": {
            "title": "Eventid",
            "type": "string"
          },
          "isDeleted": {
            "default": false,
            "title": "Isdeleted",
            "type": "boolean"
          }
        },
        "required": [
          "eventId",
          "determination"
        ],
        "title": "EventComment",
        "type": "object"
      },
      "EventTag": {
        "properties": {
          "eventId": {
            "title": "Eventid",
            "type": "string"
          },
          "isDeleted": {
            "default": false,
            "title": "Isdeleted",
            "type": "boolean"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "tag"
        ],
        "title": "EventTag",
        "type": "object"
      },
...
      "KustoQueryEcho": {
        "description": "Echo of the request; `requestedBy` is the token identity (read-only).",
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "database": {
            "title": "Database",
            "type": "string"
          },
          "endTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endtime"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "requestedBy": {
            "title": "Requestedby",
            "type": "string"
          },
          "startTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starttime"
          }
        },
        "required": [
          "cluster",
          "database",
          "query",
          "requestedBy"
        ],
        "title": "KustoQueryEcho",
        "type": "object"
      },
      "KustoQueryRequest": {
        "description": "Body of `POST /api/kusto/query`. `requestedBy` is ignored (owner comes from the token).",
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "database": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Database",
            "type": "string"
          },
          "endTime": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endtime"
          },
          "query": {
            "maxLength": 100000,
            "title": "Query",
            "type": "string"
          },
          "startTime": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starttime"
          }
        },
        "required": [
          "cluster",
          "database",
          "query"
        ],
        "title": "KustoQueryRequest",
        "type": "object"
      },
      "KustoQueryRun": {
        "description": "Run resource returned by POST (200/202) and GET. Server-produced only, never a request.",
        "properties": {
          "executeDateTimeUtc": {
            "title": "Executedatetimeutc",
            "type": "string"
          },
          "executionMetrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KustoQueryStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "expiresAt": {
            "title": "Expiresat",
            "type": "string"
          },
          "kustoQuery": {
            "$ref": "#/components/schemas/KustoQueryEcho"
          },
          "mainError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mainerror"
          },
          "queryRunId": {
            "format": "uuid",
            "title": "Queryrunid",
            "type": "string"
          },
          "resultData": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultdata"
          },
          "status": {
            "$ref": "#/components/schemas/QueryRunStatus"
          }
        },
        "required": [
          "queryRunId",
          "status",
          "kustoQuery",
          "executeDateTimeUtc",
          "expiresAt"
        ],
        "title": "KustoQueryRun",
        "type": "object"
      },
      "KustoQueryStats": {
        "additionalProperties": true,
        "description": "Kusto `QueryCompletionInformation` stats. Keys are snake_case as Kusto returns them.",
        "properties": {
          "dataset_statistics": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Statistics"
          },
          "execution_time": {
            "title": "Execution Time",
            "type": "number"
          },
          "input_dataset_statistics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Dataset Statistics"
          },
          "resource_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Usage"
          }
        },
        "required": [
          "execution_time"
        ],
        "title": "KustoQueryStats",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "The one error body (RFC 7807 style).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "traceId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "QueryRunStatus": {
        "enum": [
          "created",
          "completed",
          "error",
          "timedOut"
        ],
        "title": "QueryRunStatus",
        "type": "string"
      },
...
      "SavedEvent": {
        "properties": {
          "eventAsJson": {
            "additionalProperties": true,
            "title": "Eventasjson",
            "type": "object"
          },
          "eventId": {
            "title": "Eventid",
            "type": "string"
          },
          "eventTime": {
            "format": "date-time",
            "title": "Eventtime",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "eventTime",
          "eventAsJson"
        ],
        "title": "SavedEvent",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TIM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/kusto/query": {
      "post": {
        "operationId": "runQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KustoQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KustoQueryRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KustoQueryRun"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Consent required or Kusto denied the user (`consent-required` / `forbidden`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Upstream failure (`upstream`)."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Dependency unavailable (`unavailable`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Start Query",
        "tags": [
          "kusto"
        ]
      }
    },
    "/api/kusto/query/{queryRunId}": {
      "get": {
        "operationId": "getQueryRun",
        "parameters": [
          {
            "in": "path",
            "name": "queryRunId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queryrunid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KustoQueryRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found (`not-found`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Query Run",
        "tags": [
          "kusto"
        ]
      }
    },
...
    "/api/taggedevents/comments": {
      "post": {
        "operationId": "commentEvents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/EventComment"
                },
                "maxItems": 1000,
                "minItems": 1,
                "title": "Items",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Upstream failure (`upstream`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Add Comments",
        "tags": [
          "taggedevents"
        ]
      }
    },
    "/api/taggedevents/savedEvents": {
      "post": {
        "operationId": "saveEvents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/SavedEvent"
                },
                "maxItems": 1000,
                "minItems": 1,
                "title": "Items",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Upstream failure (`upstream`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Save Events",
        "tags": [
          "taggedevents"
        ]
      }
    },
    "/api/taggedevents/tags": {
      "post": {
        "operationId": "tagEvents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/EventTag"
                },
                "maxItems": 1000,
                "minItems": 1,
                "title": "Items",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Upstream failure (`upstream`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Add Tags",
        "tags": [
          "taggedevents"
        ]
      }
    },
...
  }
}